Added BillyJoel.py
Background Program that checks the time periodically. It plays Piano Man by Billy Joel at 9:00 PM on a Saturday. You have to specify the filepath in the code, and nab your own .mp3 of the song because of copyright. Feel free to use this however you wish, and please improve it!